Refactored multiple responsibilities of FetchBotStatus into two seperate objects
