+++ lib/command_objects/commands.rb
@@ -7,6 +7,7 @@ require_relative 'create_step'
 require_relative 'create_sequence'
 require_relative 'create_schedule'
 require_relative 'create_step'
+require_relative 'report_bot_status'
 require_relative 'fetch_bot_status'
 require_relative 'if_statement'
 require_relative 'resolve_controller'
