+++ lib/controllers/read_status_controller.rb
@@ -1,9 +1,9 @@
 require_relative 'abstract_controller'
-require_relative '../command_objects/fetch_bot_status'
+require_relative '../command_objects/report_bot_status'
 module FBPi
   class ReadStatusController < AbstractController
     def call
-      reply "read_status", FetchBotStatus.run!(bot: bot)
+      reply "read_status", ReportBotStatus.run!(bot: bot)
     end
   end
 end
